interactive models updated with wagtail serve method
