+++ iogt/templates/generic_components/interactive_card.html
@@ -1,9 +1,14 @@
 {% load static wagtailcore_tags wagtailimages_tags sass_tags image_tags home_tags %}
 
 <div class="section-card">
-    <a href="{% pageurl interactive %}" class="sub-section " style="color:#444; background:#ffffff;">
-        <div class="section-header">
-            <p class="section-title">{{ interactive.title }}</p>
-        </div>
-    </a>
+    <form id="myForm" action="{% pageurl interactive %}" method="post">
+        {% csrf_token %}
+        {{ form.as_p }}
+        <input type="hidden" name="text" value="{{ interactive.trigger_string }}">
+        <button type="submit" class="sub-section" style="color:#444; background:#ffffff; border:none; padding:0;">
+            <div class="section-header">
+                <p class="section-title" style="font-family: Poppins,serif">{{ interactive.title }}</p>
+            </div>
+        </button>
+    </form>
 </div>
