+++ iogt/urls.py
@@ -70,7 +70,6 @@ urlpatterns = api_url_patterns + [
     path('page-tree/<int:page_id>/', PageTreeAPIView.as_view(), name='page_tree'),
     path('api/docs/', schema_view.with_ui('swagger'), name='swagger'),
     path('webpush/subscribe/', save_info, name='save_webpush_info'),
-    *i18n_patterns(path('interactive/', include('interactive.urls', namespace='interactive'))),
 ]
 
 if settings.DEBUG:
